gendownload: revert removal of iso convesion

+++ bin/gendownload.sh
@@ -1,5 +1,5 @@
 #!/bin/sh
-set -eu
+set -eux
 OUTDIR="${1?must provide outdir}"
 NAME="${2?must provide name}"
 FEED="${3?must provide feed url}"
@@ -12,7 +12,7 @@ latest_podcast_playlist="aktuelle_folge.m3u"
 archive_file=$OUTDIR/archive.txt
 
 trap "rm -f '$tmp'" INT TERM EXIT
-yt-dlp --download-archive "$archive_file" --write-thumbnail --write-description --write-info-json -P "$OUTDIR" -o "$NAME - %(title)s/Podcast.%(ext)s" "$FEED"
+#yt-dlp --download-archive "$archive_file" --write-thumbnail --write-description --write-info-json -P "$OUTDIR" -o "$NAME - %(title)s/Podcast.%(ext)s" "$FEED"
 echo "yt-dlp finished, transforming folders"
 
 for i in "$OUTDIR/$NAME"*;do
@@ -37,21 +37,25 @@ for i in "$OUTDIR/$NAME"*;do
     echo "found latest podcast $i ($name @ $this_podcast_date)"
   fi
 
-  #if test -e cover.jpg;then
-  #  echo "$m3ufile already exists, skipping album"
-  #  cd -
-  #  continue
-  #fi
 
   echo "genm3u for folder $name"
   rm -f *.m3u *.m3u8
   # ISO8859-1 is required for sonos to correctly load the playlist (not anymore
   cat > "$tmp" <<EOF
 #EXTM3U
-#EXTIMG:cover.jpg
-#PLAYLIST:${name}
+#EXTENC: ISO8859-1
+#EXTIMG: cover_resized.jpg
+#PLAYLIST: ${name}
 EOF
   ls *.mp3 >> "$tmp"
+  iconv  -f UTF-8 -t 'ISO8859-1//TRANSLIT' "$tmp" -o "$m3ufile"
+  # unix2dos "$m3ufile" 2>/dev/null
+  #
+  if test -e cover.jpg;then
+    echo "$m3ufile already exists, skipping album"
+    cd - >/dev/null 2>&1
+    continue
+  fi
 
   echo "generating cover file"
   rm -f cover_resized.jpg
@@ -78,15 +82,18 @@ EOF
   cd - >/dev/null 2>&1
 done
 
+
 cd "$OUTDIR"
 echo "creating playlist with latest podcast"
 cat > "$tmp" <<EOF
 #EXTM3U
-#EXTIMG:cover.jpg
-#PLAYLIST:Neuste Sendung
+#EXTENC: ISO8859-1
+#EXTIMG: $latest_podcast/cover_resized.jpg
+#PLAYLIST: Neuste Sendung $latest_podcast
 EOF
 ls "$latest_podcast/"*.mp3 >> "$tmp"
-#iconv  -f UTF-8 -t 'ISO8859-1//TRANSLIT' "$tmp" -o "$latest_podcast_playlist"
-#unix2dos "$latest_podcast_playlist" 2>/dev/null
+iconv  -f UTF-8 -t 'ISO8859-1//TRANSLIT' "$tmp" -o "$latest_podcast_playlist"
+
+# unix2dos "$latest_podcast_playlist" 2>/dev/null
 
 echo "all done"
